Rename some files for multiple file type support
kmod vs pe still needs a bit of work to be full featured on the kmod
side, including a big refactoring, but this is a good start to making it
more generic.
